Fixes for win32

+++ src/2aes/aes.h
@@ -48,19 +48,19 @@ class AesCipher {
     AesCipher(const uint8_t key[kAesBlockSize], uint64_t salt = 0) {
       uint64_t iv[2] = {salt, 0};
 
-	  EVP_CIPHER_CTX_init(&m_encrypt_ctx);
-	  EVP_EncryptInit_ex(&m_encrypt_ctx, EVP_aes_128_cbc(), NULL, key,
-						(uint8_t *)&iv[0]);
-	  EVP_CIPHER_CTX_init(&m_decrypt_ctx);
-	  EVP_DecryptInit_ex(&m_decrypt_ctx, EVP_aes_128_cbc(), NULL, key,
-						(uint8_t *)&iv[0]);
-	  EVP_CIPHER_CTX_set_padding(&m_encrypt_ctx, 0);
-	  EVP_CIPHER_CTX_set_padding(&m_decrypt_ctx, 0);
+	    EVP_CIPHER_CTX_init(&m_encrypt_ctx);
+	    EVP_EncryptInit_ex(&m_encrypt_ctx, EVP_aes_128_cbc(), NULL, key,
+			  			(uint8_t *)&iv[0]);
+	    EVP_CIPHER_CTX_init(&m_decrypt_ctx);
+	    EVP_DecryptInit_ex(&m_decrypt_ctx, EVP_aes_128_cbc(), NULL, key,
+			  			(uint8_t *)&iv[0]);
+	    EVP_CIPHER_CTX_set_padding(&m_encrypt_ctx, 0);
+	    EVP_CIPHER_CTX_set_padding(&m_decrypt_ctx, 0);
     }
 
     ~AesCipher() {
       EVP_CIPHER_CTX_cleanup(&m_encrypt_ctx);
-	  EVP_CIPHER_CTX_cleanup(&m_decrypt_ctx);
+  	  EVP_CIPHER_CTX_cleanup(&m_decrypt_ctx);
     }
 
     /*
@@ -69,17 +69,17 @@ class AesCipher {
      *
      * The input data length must be aligned to the aes block size (16 bytes)!
      */
-    void encrypt(const uint8_t *plaintext, uint8_t *ciphertext, int len) {
-	  assert(len % kAesBlockSize == 0);
+    void encrypt(const uint8_t *plaintext, uint8_t *ciphertext, size_t len) {
+	    assert(len % kAesBlockSize == 0);
 
-	  /* update ciphertext, c_len is filled with the length of ciphertext
-	   * generated, len is the size of plaintext in bytes */
-	  int clen = len;
-	  EVP_EncryptUpdate(&m_encrypt_ctx, ciphertext, &clen, plaintext, len);
+	    /* update ciphertext, c_len is filled with the length of ciphertext
+	     * generated, len is the size of plaintext in bytes */
+	    int clen = (int)len;
+	    EVP_EncryptUpdate(&m_encrypt_ctx, ciphertext, &clen, plaintext, (int)len);
 
-	  /* update ciphertext with the final remaining bytes */
-	  int outlen;
-	  EVP_EncryptFinal(&m_encrypt_ctx, ciphertext + clen, &outlen);
+	    /* update ciphertext with the final remaining bytes */
+	    int outlen;
+	    EVP_EncryptFinal(&m_encrypt_ctx, ciphertext + clen, &outlen);
     }
 
     /*
@@ -88,13 +88,12 @@ class AesCipher {
      *
      * The input data length must be aligned to the aes block size (16 bytes)!
      */
-    void decrypt(const uint8_t *ciphertext, uint8_t *plaintext, int len) {
+    void decrypt(const uint8_t *ciphertext, uint8_t *plaintext, size_t len) {
       assert(len % kAesBlockSize == 0);
 
-	  int plen = len, flen = 0;
-
-	  EVP_DecryptUpdate(&m_decrypt_ctx, plaintext, &plen, ciphertext, len);
-	  EVP_DecryptFinal(&m_decrypt_ctx, plaintext + plen, &flen);
+	    int plen = (int)len, flen = 0;
+	    EVP_DecryptUpdate(&m_decrypt_ctx, plaintext, &plen, ciphertext, (int)len);
+	    EVP_DecryptFinal(&m_decrypt_ctx, plaintext + plen, &flen);
     }
 
   private:
